style: move test-only code under `_test` files
