+++ .github/workflows/test.yml
@@ -47,6 +47,11 @@ jobs:
         run: |
           set -x
           go vet ./...
+      - name: Check for unused code
+        run: |
+          set -x
+          go install honnef.co/go/tools/cmd/staticcheck@latest
+          make unused
       - name: Check gosec
         run: |
           set -x
