+++ api/helpers_test.go
@@ -1,0 +1,23 @@
+package api
+
+import (
+	"net"
+)
+
+func removeLocalhostFromPrivateIPBlock() *net.IPNet {
+	_, localhost, _ := net.ParseCIDR("127.0.0.0/8")
+
+	var localhostIndex int
+	for i := 0; i < len(privateIPBlocks); i++ {
+		if privateIPBlocks[i] == localhost {
+			localhostIndex = i
+		}
+	}
+	privateIPBlocks = append(privateIPBlocks[:localhostIndex], privateIPBlocks[localhostIndex+1:]...)
+
+	return localhost
+}
+
+func unshiftPrivateIPBlock(address *net.IPNet) {
+	privateIPBlocks = append([]*net.IPNet{address}, privateIPBlocks...)
+}
